+++ src/index.js
@@ -2,4 +2,6 @@ import Input from './input';
 import Icon from './icon';
 import Button from './button';
 
+export * from './icon-button';
+
 export { Input, Button, Icon };