+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.0.7 [08/10/2020]
+
+-   **Created** the `IconButton` component
+-   **Created** the `Tooltip` component
+
 ## 0.0.1 [unreleased]
 
 ### 🦄 New Features 🦄